MTDSA-12900 update connector spec
